v1.71: C safety fixes, library API, and TypeScript/WASM scaffold
C cleanup (Phase 1):
- PW_in_size and PW_Start_Address widened from int32_t to int64_t;
  PWGetFileSize() return type updated to match
- strcpy → strncpy for OutName_final in Save_Rip/Save_Rip_Special
- NULL check added in Save_Rip_Special before fwrite
- PW_FindResult struct added to globals.h for structured result passing
- PW_RecordFind() extracted as pure-logic function (no I/O); Save_Rip
  and Save_Rip_Special refactored to call it
- Scan loop extracted from main() into PW_ScanData() so library code
  can invoke the engine without spawning a process
- Format-string warnings fixed in r/DoubleAction.c, r/IT.c, r/SA.c

New:
- prowiz_lib.c: prowiz_scan() (returns malloc'd JSON) and
  prowiz_extract() (raw byte copy) for WASM / library use
- Makefile: wasm target for Emscripten build
- packages/prowiz-wasm/: TypeScript types, WASM loader, public API
  (createProwiz / scan / extract / identify)

+++ r/DoubleAction.c
@@ -94,7 +94,7 @@ void Rip_DoubleAction10 ( void )
 
 /*     Save_Rip_Special ( "Double Action 1.0 Exe-file", Double_Action, Amiga_EXE_Header_Block , 24 ); */
 /*     free ( Amiga_EXE_Header_Block ); */
-    printf ("Double Action 1.0 Exe-file without header detected at %d. Can't rebuild it, sorry\n", PW_Start_Address);
+    printf ("Double Action 1.0 Exe-file without header detected at %lld. Can't rebuild it, sorry\n", (long long)PW_Start_Address);
   }
   else
   {
